+++ src/style/style.css
@@ -100,13 +100,11 @@ header{
 }
 
 #staffContainer {
-   padding: 170px;
+   padding: 200px;
 }
 
 #contactContainer h1 {
-   position: relative;
    padding: 250px 0px 0px 0px;
-
 }
 
 #contactContainer h3{
@@ -218,9 +216,9 @@ footer {
         width: 100%;
         opacity: 1;
         text-align: center;
-     }
+    }
      
-     #welcome-message{
+    #welcome-message{
          margin: 20px 0px;
          position: relative;
          padding: 0px 20px;
@@ -228,7 +226,7 @@ footer {
          font-size: 2rem;
          font-family: 'Merriweather', serif;
          color: #000;
-     }
+    }
      
     #short-description {
         margin: 30px 0px;
@@ -238,16 +236,16 @@ footer {
         color: rgb(25, 36, 25);
     }
 
-     #menuContainer h1 {
+    #menuContainer h1 {
         padding: 330px 0px 0px 0px;
-     }
+    }
      
-     #menuContainer ul {
+    #menuContainer ul {
          padding: 20px; 
          list-style: none;
-     }
+    }
      
-     #menuContainer li{
+    #menuContainer li{
         margin: 10px auto;
         padding: 12px 30px;
         width: 50%;
@@ -256,61 +254,60 @@ footer {
         color: #000;
         font-weight: 700;
         border-radius: 5px;
-     }
+    }
      
-     #staffContainer {
+    #staffContainer {
        margin-top: 0px;
        padding-top: 300px;
-     }
+    }
      
-     #staffContainer h1 {
+    #staffContainer h1 {
         display: none;
         margin: 0px;
         padding: 0px;
-     }
+    }
 
-     #staffContainer p {
+    #staffContainer p {
         margin: 0px;
         padding: 0px;
         display: block;
-     }
+    }
 
-     #contactContainer h1 {
-        position: relative;
+    #contactContainer h1 {
         padding: 340px 0px 0px 0px;
-     }
+    }
      
-     #contactContainer h3{
-         position: relative;
-         font-weight: bold;
-         font-size: 1.5rem;
-         color: black;
-         font-family: 'Merriweather', serif;
-     }
+    #contactContainer h3{
+        position: relative;
+        font-weight: bold;
+        font-size: 1.5rem;
+        color: black;
+        font-family: 'Merriweather', serif;
+    }
      
-     #contact-details {
+    #contact-details {
          padding: 0px 0px;
-     }
+    }
      
-     footer {
-         width: 100%;
-         height: 50px;
-         position: fixed;
-         bottom: 0;
-         background-color: rgb(107, 5, 5);
-         color: white;
-     }
+    footer {
+        width: 100%;
+        height: 50px;
+        position: fixed;
+        bottom: 0;
+        background-color: rgb(107, 5, 5);
+        color: white;
+    }
      
-     #auth {
-         position: relative;
-         top: 25%;
-         -webkit-transform: translateY(-50%);
-         -ms-transform: translateY(-50%);
-         transform: translateY(-50%);
-         text-align: center;
-         font-weight: bold;
-         font-size: small;
-     }
+    #auth {
+        position: relative;
+        top: 25%;
+        -webkit-transform: translateY(-50%);
+        -ms-transform: translateY(-50%);
+        transform: translateY(-50%);
+        text-align: center;
+        font-weight: bold;
+        font-size: small;
+    }
 }
   
 @media (min-width: 576px) and (max-width: 992px) {
@@ -380,14 +377,14 @@ footer {
         text-align: center;
      }
      
-     #welcome-message{
-         margin: 20px 0px;
-         position: relative;
-         padding: 0px 20px;
-         font-weight: bold;
-         font-size: 2rem;
-         font-family: 'Merriweather', serif;
-         color: #000;
+    #welcome-message{
+        margin: 20px 0px;
+        position: relative;
+        padding: 0px 20px;
+        font-weight: bold;
+        font-size: 2rem;
+        font-family: 'Merriweather', serif;
+        color: #000;
      }
      
     #short-description {
@@ -398,16 +395,16 @@ footer {
         color: rgb(25, 36, 25);
     }
 
-     #menuContainer h1 {
+    #menuContainer h1 {
         padding: 130px 0px 0px 0px;
-     }
+    }
      
-     #menuContainer ul {
-         padding: 20px; 
-         list-style: none;
-     }
-     
-     #menuContainer li{
+    #menuContainer ul {
+        padding: 20px; 
+        list-style: none;
+    }
+    
+    #menuContainer li{
         margin: 10px auto;
         padding: 12px 30px;
         width: 50%;
@@ -416,58 +413,56 @@ footer {
         color: #000;
         font-weight: 700;
         border-radius: 5px;
-     }
+    }
      
-     #staffContainer {
+    #staffContainer {
        margin-top: 0px;
        padding-top: 130px;
-     }
+    }
      
-     #staffContainer h1 {
+    #staffContainer h1 {
         margin: 0px;
         padding: 0px;
-     }
+    }
 
-     #staffContainer p {
+    #staffContainer p {
         margin: 0px;
         padding: 0px;
         display: block;
-     }
+    }
 
-     #contactContainer h1 {
-        position: relative;
+    #contactContainer h1 {
         padding: 140px 0px 0px 0px;
+    }
+     
+    #contactContainer h3{
+        font-weight: bold;
+        font-size: 1.5rem;
+        color: black;
+        font-family: 'Merriweather', serif;
+    }
+     
+    #contact-details {
+        padding: 0px 0px;
+    }
+     
+    footer {
+        width: 100%;
+        height: 50px;
+        position: fixed;
+        bottom: 0;
+        background-color: rgb(107, 5, 5);
+        color: white;
      }
      
-     #contactContainer h3{
-         position: relative;
-         font-weight: bold;
-         font-size: 1.5rem;
-         color: black;
-         font-family: 'Merriweather', serif;
-     }
-     
-     #contact-details {
-         padding: 0px 0px;
-     }
-     
-     footer {
-         width: 100%;
-         height: 50px;
-         position: fixed;
-         bottom: 0;
-         background-color: rgb(107, 5, 5);
-         color: white;
-     }
-     
-     #auth {
-         position: relative;
-         top: 25%;
-         -webkit-transform: translateY(-50%);
-         -ms-transform: translateY(-50%);
-         transform: translateY(-50%);
-         text-align: center;
-         font-weight: bold;
-         font-size: small;
-     }
+    #auth {
+        position: relative;
+        top: 25%;
+        -webkit-transform: translateY(-50%);
+        -ms-transform: translateY(-50%);
+        transform: translateY(-50%);
+        text-align: center;
+        font-weight: bold;
+        font-size: small;
+    }
 }
